Address PR review: move resolved bindings to experimental, fix pointer stability
- Create experimental/descriptor_patching.hpp with ResolvedRtArgBinding,
  ResolvedCbBinding, ResolvedBindings, resolve_bindings, apply_resolved_bindings,
  moving them out of the stable program_descriptors.hpp per akerteszTT's request.
  This shim is explicitly marked for removal once ops migrate to Metal 2.0.

- Fix apply_resolved_bindings to re-derive RuntimeArgsData via GetRuntimeArgs /
  GetCommonRuntimeArgs on every cache-hit call instead of storing a raw
  RuntimeArgsData* captured at cache-miss time.  The dispatch path mutates
  rt_args_data inside the struct on first enqueue (retargeting it to the
  command-sequence buffer); re-deriving via the stable struct lookup is safe
  across that boundary without any cross-call pointer state.
  ResolvedRtArgBinding now stores (kernel_idx, core, arg_idx, is_common)
  instead of RuntimeArgsData*.

- Remove runtime_args_data.hpp from program_descriptors.hpp includes (no longer
  referenced in that header).

- Fix Clang-Tidy: RTArgList overloads of emplace_runtime_args /
  emplace_common_runtime_args now take const RTArgList& instead of by value.

Made-with: Cursor
