+++ tt_metal/api/tt-metalium/program_descriptors.hpp
@@ -10,7 +10,6 @@
 #include <tt-metalium/circular_buffer_constants.h>
 #include <tt-metalium/kernel_types.hpp>
 #include <tt-metalium/mesh_coord.hpp>
-#include <tt-metalium/runtime_args_data.hpp>
 #include <tt_stl/small_vector.hpp>
 
 // UMD: re-exports CoreType (used in SemaphoreDescriptor::core_type member).
@@ -176,13 +175,15 @@ struct KernelDescriptor {
     // as buffer bindings at their position.  Use this instead of
     // runtime_args.emplace_back() when some args are buffer base addresses.
     void emplace_runtime_args(const CoreCoord& core, std::initializer_list<std::variant<uint32_t, Buffer*>> args);
-    void emplace_runtime_args(const CoreCoord& core, RTArgList args);
+    void emplace_runtime_args(const CoreCoord& core, const RTArgList& args);
+    // Vector overload for dynamically-built arg lists.
+    void emplace_runtime_args(const CoreCoord& core, std::vector<std::variant<uint32_t, Buffer*>> args);
 
     // Push common runtime args, automatically registering any Buffer* entries
     // as common buffer bindings.  Use this instead of assigning common_runtime_args
     // directly when some args are buffer base addresses.
     void emplace_common_runtime_args(std::initializer_list<std::variant<uint32_t, Buffer*>> args);
-    void emplace_common_runtime_args(RTArgList args);
+    void emplace_common_runtime_args(const RTArgList& args);
 };
 
 struct ProgramDescriptor {
@@ -217,55 +218,6 @@ ProgramDescriptor merge_program_descriptors(const std::vector<ProgramDescriptor>
  * descriptor structure.
  */
 void apply_descriptor_runtime_args(Program& program, const ProgramDescriptor& desc);
-
-// ----------------------------------------------------------------------------
-// Fast cache-hit patching support
-// ----------------------------------------------------------------------------
-
-// A buffer binding resolved to a direct pointer into a cached Program's
-// runtime args storage.  Created once on cache miss; used on every cache hit.
-// tensor_buffer_idx indexes into the vector returned by collect_tensor_buffers()
-// at both resolve time and apply time — no raw Buffer* stored here.
-struct ResolvedRtArgBinding {
-    RuntimeArgsData* data = nullptr;
-    uint32_t arg_idx = 0;
-    uint32_t tensor_buffer_idx = 0;
-};
-
-// A CB dynamic-address binding resolved to a CB id for direct update.
-// cb_id is CBHandle (uintptr_t) — a 64-bit handle, not an index.
-// tensor_buffer_idx indexes into the same collect_tensor_buffers() vector.
-struct ResolvedCbBinding {
-    uintptr_t cb_id = 0;
-    uint32_t tensor_buffer_idx = 0;
-    uint32_t address_offset = 0;
-};
-
-// All resolved bindings for one program.  Non-empty only when the factory
-// declared at least one buffer arg via emplace_runtime_args().
-struct ResolvedBindings {
-    std::vector<ResolvedRtArgBinding> rt_args;
-    std::vector<ResolvedCbBinding> cbs;
-    bool empty() const noexcept { return rt_args.empty() && cbs.empty(); }
-};
-
-// Walk desc.kernels[k].buffer_bindings and desc.cbs[i].buffer, resolve each
-// to a stable pointer/id into the already-built program, and return the result.
-// tensor_buffers is an ordered enumeration of all Buffer* reachable from the
-// current call's tensor_args and tensor_return_value (built by the adapter via
-// collect_tensor_buffers).  Every binding Buffer* must appear in tensor_buffers;
-// the matching index is stored in place of the raw pointer.
-// Call immediately after Program{desc} on cache miss; store in shared_variables.
-ResolvedBindings resolve_bindings(
-    Program& program, const ProgramDescriptor& desc, const std::vector<Buffer*>& tensor_buffers);
-
-// Apply resolved bindings to the cached program on a cache hit.
-// current_buffers must be the output of collect_tensor_buffers() for the
-// current call's tensors — same enumeration order as at resolve time.
-// Writes buffer->address() directly into the program's runtime args storage
-// and calls UpdateDynamicCircularBufferAddress for any dynamic CBs.
-void apply_resolved_bindings(
-    Program& program, const ResolvedBindings& bindings, const std::vector<Buffer*>& current_buffers);
 
 }  // namespace tt::tt_metal
